refactor(routes): add comments and add authorization middleware on specific routes
