+++ src/routes/userRoute.js
@@ -1,9 +1,28 @@
 const userCtrl = require('../controllers/userCtrl');
-const { authMiddleware } = require('../middleware/authMiddleware');
+
+const { authMiddleware } = require('../middleware/auth.middleware');
+const { isAuthorized } = require('../middleware/authorization.middleware');
+const { ROLES } = require('../config/global');
 
 module.exports = (app) => {
-  app.get('/users', authMiddleware, userCtrl.getUsers);
+  // Get single user by id
   app.get('/user/:userId', authMiddleware, userCtrl.getOneUser);
-  app.delete('/user/:userId', authMiddleware, userCtrl.deleteUser);
+
+  // Update single user by id
   app.put('/user/:userId', authMiddleware, userCtrl.updateUser);
+
+  // Remove single user by id
+  app.delete('/user/:userId', authMiddleware, userCtrl.deleteUser);
+
+  /**
+   * Admin routes
+   */
+
+  // Get all users
+  app.get(
+    '/admin/users',
+    authMiddleware,
+    isAuthorized(ROLES.admin),
+    userCtrl.getUsers
+  );
 };
